move _sym_inv to utils.linalg
